sstables: mc: sign-extend serialization_header min_local_deletion_time_base and min_ttl_base
Refs #4074
Refs #3353

Message-Id: <20190110141439.1324-1-bhalevy@scylladb.com>
(cherry picked from commit 2dc377640781e8e7e1d969cfeccf4b1380218ebd)

+++ sstables/mc/writer.cc
@@ -367,9 +367,9 @@ static sstring pk_type_to_string(const schema& s) {
 serialization_header make_serialization_header(const schema& s, const encoding_stats& enc_stats) {
     serialization_header header;
     // mc serialization header minimum values are delta-encoded based on the default timestamp epoch times
-    header.min_timestamp_base.value = static_cast<uint64_t>(enc_stats.min_timestamp) - encoding_stats::timestamp_epoch;
-    header.min_local_deletion_time_base.value = static_cast<uint32_t>(enc_stats.min_local_deletion_time) - encoding_stats::deletion_time_epoch;
-    header.min_ttl_base.value = static_cast<uint32_t>(enc_stats.min_ttl) - encoding_stats::ttl_epoch;
+    header.min_timestamp_base.value = static_cast<uint64_t>(enc_stats.min_timestamp - encoding_stats::timestamp_epoch);
+    header.min_local_deletion_time_base.value = static_cast<uint64_t>(enc_stats.min_local_deletion_time - encoding_stats::deletion_time_epoch);
+    header.min_ttl_base.value = static_cast<uint64_t>(enc_stats.min_ttl - encoding_stats::ttl_epoch);
 
     header.pk_type_name = to_bytes_array_vint_size(pk_type_to_string(s));
